requirements.txt
numpy>1.16
pip3
scipy
matplotlib
pytest
setuptools>=26
autopep8
pandas
scikit-learn
seaborn
jupyter
jupyterlab
pytest
pytz
bs4
datetime
PyExecJS
requests
re
h5py
pytz
